requirements.txt
piexif
Pillow
pillow-heif